feat: rename queues
