+++ packages/indexer/src/jobs/metadata-index/metadata-fetch-job.ts
@@ -9,7 +9,7 @@ import { PendingRefreshTokensBySlug } from "@/models/pending-refresh-tokens-by-s
 import { AddressZero } from "@ethersproject/constants";
 import { metadataIndexProcessJob } from "@/jobs/metadata-index/metadata-process-job";
 import { metadataIndexProcessBySlugJob } from "@/jobs/metadata-index/metadata-process-by-slug-job";
-import { onchainMetadataFetchTokenUriJob } from "./onchain-metadata-process-token-uri-job";
+import { onchainMetadataFetchTokenUriJob } from "./onchain-metadata-fetch-token-uri-job";
 
 export type MetadataIndexFetchJobPayload =
   | {
